requirements.txt
accelerate==1.10.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
bidict==0.23.1
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
controlnet-aux==0.0.3
cycler==0.12.1
diffusers==0.27.2
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.108.0
ffmpy==0.6.1
filelock==3.19.1
Flask==2.2.3
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flaskwebgui==0.3.5
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
gradio==4.21.0
gradio_client==0.12.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
IOPaint==1.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lama-cleaner==1.2.5
lazy_loader==0.4
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.1.2
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.22.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
orjson==3.11.2
packaging==25.0
pandas==2.3.1
peft==0.7.1
pgvector==0.4.1
piexif==1.1.3
Pillow==9.5.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyMatting==1.1.14
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.7.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
rembg==2.0.67
requests==2.32.5
rich==14.1.0
rpds-py==0.27.0
ruff==0.12.9
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.32.0.post1
sympy==1.14.0
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.21.4
tomlkit==0.12.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.48.3
typer==0.16.1
typer-config==1.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.181
ultralytics-thop==2.0.15
urllib3==2.5.0
uvicorn==0.35.0
websockets==11.0.3
Werkzeug==2.2.2
whichcraft==0.6.1
wsproto==1.2.0
yacs==0.1.8
zipp==3.23.0
